requirements.txt
# ===== CORE DEPENDENCIES =====
python-telegram-bot==20.7
telethon==1.34.0
asyncio==3.4.3
nest-asyncio==1.5.8

# ===== DATABASE =====
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# ===== REDIS E QUEUE =====
redis==5.0.1
rq==1.15.1
rq-scheduler==0.13.1

# ===== HTTP E API =====
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# ===== SERIALIZAÇÃO =====
msgpack==1.0.7
python-dotenv==1.0.0

# ===== PAGAMENTOS =====
mercadopago==2.2.1

# ===== MONITORAMENTO =====
prometheus-client==0.19.0
psutil==5.9.6

# ===== LOGS =====
python-json-logger==2.0.7

# ===== UTILITÁRIOS =====
pathlib2==2.3.7
typing-extensions==4.8.0
